console: Add mac settings download dropdown
